+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.2.0
+
+- Add `isActive` getter
+
 ## 0.1.2
 
-- Return ValueStream from PausableBehaviorSubject.stream
+- Return `ValueStream` from `PausableBehaviorSubject.stream`
 
 ## 0.1.1